gen-ai/9-langgraph-checkpoint/practical/requirements.txt
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.7
distro==1.9.0
dnspython==2.8.0
greenlet==3.4.0
groq==0.37.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jsonpatch==1.33
jsonpath-ng==1.8.0
jsonpointer==3.1.1
langchain==1.2.15
langchain-classic==1.0.3
langchain-core==1.2.28
langchain-groq==1.1.2
langchain-mongodb==0.11.0
langchain-text-splitters==1.1.1
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-checkpoint-redis==0.4.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.13
langsmith==0.7.30
lark==1.3.1
ml_dtypes==0.5.4
numpy==2.4.4
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.0
pydantic==2.13.0
pydantic_core==2.46.0
pymongo-search-utils==0.3.0
python-dotenv==1.2.2
python-ulid==3.1.0
PyYAML==6.0.3
redis==7.4.0
redisvl==0.17.0
requests==2.33.1
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.49
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
xxhash==3.6.0
zstandard==0.25.0
